Moved MaybeModule and dependents into separate files.
